Divided landing pages into child and parent
Removed functionality from child landing page
