+++ foodplanner/lib/pages/landing_page_children_madpakke.dart
@@ -8,7 +8,8 @@ import 'package:foodplanner/components/mealBoxEmpty.dart';
 import 'package:foodplanner/config/text_styles.dart';
 import 'package:foodplanner/pages/feedbackChatPage.dart';
 import 'package:foodplanner/components/footer.dart'; // Import the footer widget
-import 'package:intl/intl.dart'; // Import the reusable widget
+import 'package:intl/intl.dart';
+import 'package:provider/provider.dart'; // Import the reusable widget
 
 class ChildLandingPageMadpakke extends StatelessWidget {
   const ChildLandingPageMadpakke({super.key});
@@ -18,43 +19,33 @@ class ChildLandingPageMadpakke extends StatelessWidget {
     // Get the size of the screen
     final size = MediaQuery.of(context).size;
 
-    final bool isMadpakkeEmpty = false; // Later we want to check with a fetch whether there is a box or not
-
     return Scaffold(
       appBar: AppBar(),
-      body: Column(
-        children: [
-          Expanded(
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.start,
-              children: [
-                const Text(
-                  // TODO read user title and display based on who is logged in 
-                  'Welcome' + ' ' + 'Child',
-                  style: TextStyle(fontSize: 16),
+      body: Align(
+        alignment: Alignment.topCenter,
+        child: SizedBox(
+          width: size.width * 0.9, // if other pages get skewed then you can do this OR its something to do with reusableMealBox.dart
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.stretch,
+            children: [
+              Expanded(
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  mainAxisAlignment: MainAxisAlignment.start,
+                  children: [
+                    const Text(
+                      'Velkommen Forældre',
+                      style: TextStyle(fontSize: 16),
+                      textAlign: TextAlign.center,
+                    ),
+                    SizedBox(height: size.height * 0.02),
+                    ReusableMealBox(size: size),
+                  ],
                 ),
-                SizedBox(height: size.height * 0.02),
-                ReusableMealBox(size: size), // Use the reusable widget
-                SizedBox(height: size.height * 0.02),
-
-                isMadpakkeEmpty
-                    ? AddMealButton(size: size)
-                    : CustomButton(
-                  onTab: () {
-                    Navigator.push(
-                      context,
-                      MaterialPageRoute(builder: (context) => FeedbackChatPage()),
-                    );
-                  }, 
-                  text: 'Se Feedback',
-                  //fontSize: 16,
-                  customWidth: size.width * 0.6,
-                ),
-              ],
-            ),
+              ),
+            ],
           ),
-          FooterBar(), // Add the footer widget here
-        ],
+        ),
       ),
     );
   }
